requirements.txt
click==8.1.7
colorthief==0.2.1
Flask==2.0.3
fpdf==1.7.2
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.1
pillow==10.4.0
Werkzeug==2.0.3
